Add support for SMT triggers
This is implemented by allowing arbitrary attributes to be
propagated throughout the pipeline. For the SMT backend, these are
translated into SMT attributes. Currently, the only attribute supported
is the `:pattern` attribute used to indicate triggers. Adding other
attributes will be easy now that this plumbing is in place, however.

Attributes currently aren't supported for TPTP, but it should be
possible to allow them eventually if there's a use for them.

Note that the initial construction of attributes is very general, and
allows attributes of any base type supported by the translator. The set
supported by SMT is more restricted, but other prover input languages
may use entirely different attributes.

There's no attempt to ensure that attributes go in places that are
allowed. So, for instance, adding a trigger outside of a quantified
expression will lead to an error from the solver. Similarly, adding a
trigger to a goal, rather than an axiom, will typically also lead to an
error.